pyproject.toml
[project]
name = "autoseg"
version = "0.1.0"
description = "Annotation tool with SAM segmentation and propagation"
readme = "README.md"
requires-python = ">=3.12.0, <3.13"
dependencies = [
    # Core
    "numpy>=1.26.0,<2",
    "pillow>=10.0.0",
    "opencv-python>=4.8.0",
    
    # ML
    "torch>=2.7.0",
    "torchvision>=0.22.0",
    
    # FastAPI backend
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "pydantic>=2.0.0",
    
    # Data utilities
    "pycocotools>=2.0.0",
    
    # SAM3 (local submodule)
    "sam3",
    
    # SAM3 runtime dependencies (missing from sam3's base deps)
    "einops",
    "decord",
    "psutil",
    "scipy",
    
    # DINOv3 runtime dependencies (loaded via torch.hub.load)
    "torchmetrics",
    "omegaconf",
    "termcolor",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-qt>=4.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.hatch.build.targets.wheel]
packages = ["app", "core", "ml"]

[tool.uv.sources]
sam3 = { path = "third_party/sam3", editable = true }
